% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{add_class}
\alias{add_class}
\title{Convenience function to add a class}
\usage{
add_class(x, cl)
}
\arguments{
\item{x}{Object}

\item{cl}{Vector of class labels to add}
}
\description{
Convenience function to add a class
}
\examples{
foo <- "some text" \%>\% add_class("text")
foo <- "some text" \%>\% add_class(c("text", "another class"))

}
