% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{arrange_data}
\alias{arrange_data}
\title{Arrange data with user-specified expression}
\usage{
arrange_data(dataset, expr = NULL)
}
\arguments{
\item{dataset}{Data frame to arrange}

\item{expr}{Expression to use arrange rows from the specified dataset}
}
\value{
Arranged data frame
}
\description{
Arrange data with user-specified expression
}
\details{
Arrange data, likely in combination with slicing
}
