% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_factor}
\alias{as_factor}
\title{Wrapper for factor with ordered = FALSE}
\usage{
as_factor(x, ordered = FALSE)
}
\arguments{
\item{x}{Input vector}

\item{ordered}{Order factor levels (TRUE, FALSE)}
}
\description{
Wrapper for factor with ordered = FALSE
}

