% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_numeric}
\alias{as_numeric}
\title{Convert variable to numeric avoiding potential issues with factors}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
Numeric
}
\description{
Convert variable to numeric avoiding potential issues with factors
}
\examples{
as_numeric(rnorm(10))
as_numeric(letters)
as_numeric(as.factor(5:10))
as.numeric(as.factor(5:10))
as_numeric(c("a","b"))
as_numeric(c("3","4"))
as_numeric(as.factor(c("3","4")))

}
