% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{as_tibble}
\alias{as_tibble}
\title{Exporting as_tibble from tibble}
\description{
Exporting as_tibble from tibble
}
\details{
See \code{\link[tibble]{as_tibble}} in the \code{tibble} package for more details
}
