% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{center}
\alias{center}
\title{Center}
\usage{
center(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
If x is a numberic variable return x - mean(x)
}
\description{
Center
}

