% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{choose_dir}
\alias{choose_dir}
\title{Select a directory. Uses JavaScript on Mac, utils::choose.dir on Windows, and dirname(file.choose()) on Linux}
\usage{
choose_dir(...)
}
\arguments{
\item{...}{Arguments passed to  utils::choose.dir on Windows}
}
\value{
Path to the directory selected by the user
}
\description{
Select a directory. Uses JavaScript on Mac, utils::choose.dir on Windows, and dirname(file.choose()) on Linux
}
\examples{
\dontrun{
choose_dir()
}

}
