% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{choose_files}
\alias{choose_files}
\title{Choose files interactively}
\usage{
choose_files(...)
}
\arguments{
\item{...}{Strings used to indicate which file types should be available for selection (e.g., "csv" or "pdf")}
}
\value{
Vector of paths to files selected by the user
}
\description{
Choose files interactively
}
\details{
Open a file dialog. Uses JavaScript on Mac, utils::choose.files on Windows, and file.choose() on Linux
}
\examples{
\dontrun{
choose_files("pdf", "csv")
}

}
