% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{copy_attr}
\alias{copy_attr}
\title{Copy attributes from on object to another}
\usage{
copy_attr(to, from, attr)
}
\arguments{
\item{to}{Object to copy attributes to}

\item{from}{Object to copy attributes from}

\item{attr}{Vector of attributes. If missing all attributes will be copied}
}
\description{
Copy attributes from on object to another
}

