% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{deregister}
\alias{deregister}
\title{Deregister a data.frame or list in Radiant}
\usage{
deregister(dataset)
}
\arguments{
\item{dataset}{String containing the name of the data.frame to register}
}
\description{
Deregister a data.frame or list in Radiant
}
