% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{describe}
\alias{describe}
\title{Show dataset desription, if available, in html form in Rstudio viewer or default browser}
\usage{
describe(name)
}
\arguments{
\item{name}{Dataset name or a dataframe}
}
\description{
Show dataset desription, if available, in html form in Rstudio viewer or default browser
}

