% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{does_vary}
\alias{does_vary}
\title{Does a vector have non-zero variability?}
\usage{
does_vary(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Logical. TRUE is there is variability
}
\description{
Does a vector have non-zero variability?
}
\examples{
summarise_all(diamonds, funs(does_vary)) \%>\% as.logical

}
