% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{dtab.character}
\alias{dtab.character}
\title{Create a DT table with bootstrap theme}
\usage{
\method{dtab}{character}(...)
}
\arguments{
\item{...}{Arguments to pass on to dtab.data.frame}
}
\description{
Create a DT table with bootstrap theme
}
\details{
View, search, sort, etc. your data. For styling options see \url{http://rstudio.github.io/DT/functions.html}
}
\examples{
dtab("mtcars")

}
