% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{dtab}
\alias{dtab}
\title{Method to create datatables}
\usage{
dtab(object, ...)
}
\arguments{
\item{object}{Object of relevant class to render}

\item{...}{Additional arguments}
}
\description{
Method to create datatables
}
