% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{empty_level}
\alias{empty_level}
\title{Convert categorical variables to factors and deal with empty/missing values}
\usage{
empty_level(x)
}
\arguments{
\item{x}{Categorical variable used in table}
}
\value{
Variable with updated levels
}
\description{
Convert categorical variables to factors and deal with empty/missing values
}
