% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{factorizer}
\alias{factorizer}
\title{Convert character to factors as needed}
\usage{
factorizer(dataset, safx = 30)
}
\arguments{
\item{dataset}{Data frame}

\item{safx}{Values to levels ratio}
}
\value{
Data frame with factors
}
\description{
Convert character to factors as needed
}
