% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{filterdata}
\alias{filterdata}
\title{Filter data with user-specified expression}
\usage{
filterdata(dataset, filt = "", drop = TRUE)
}
\arguments{
\item{dataset}{Data frame to filter}

\item{filt}{Filter expression to apply to the specified dataset (e.g., "price > 10000" if dataset is "diamonds")}

\item{drop}{Drop unused factor levels after filtering (default is TRUE)}
}
\value{
Filtered data frame
}
\description{
Filter data with user-specified expression
}
