% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{find_dropbox}
\alias{find_dropbox}
\title{Find Dropbox folder}
\usage{
find_dropbox(account = 1)
}
\arguments{
\item{account}{Integer. If multiple accounts exist, specify which one to use. By default, the first account listed is used}
}
\value{
Path to Dropbox account
}
\description{
Find Dropbox folder
}
\details{
Find the path for Dropbox if available
}
