% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{find_gdrive}
\alias{find_gdrive}
\title{Find Google Drive folder}
\usage{
find_gdrive()
}
\value{
Path to Google Drive folder
}
\description{
Find Google Drive folder
}
\details{
Find the path for Google Drive if available
}
