% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{find_project}
\alias{find_project}
\title{Find the rstudio project directory}
\usage{
find_project(mess = TRUE)
}
\arguments{
\item{mess}{Show or hide messages (default mess = TRUE)}
}
\value{
Path to rstudio project directory
}
\description{
Find the rstudio project directory
}
