% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{fix_names}
\alias{fix_names}
\title{Make column names that are valid in R}
\usage{
fix_names(x)
}
\arguments{
\item{x}{Data.frame or vector of column names}
}
\description{
Make column names that are valid in R
}
\details{
Removes symbols, trailing and leading spaces and converts to valid R column names
}
