% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{fix_smart}
\alias{fix_smart}
\title{Replace smart quotes etc.}
\usage{
fix_smart(text, all = FALSE)
}
\arguments{
\item{text}{Text to be parsed}

\item{all}{Should all non-ascii characters be removed? Default is FALSE}
}
\description{
Replace smart quotes etc.
}
