% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{fixMS}
\alias{fixMS}
\title{Replace Windows smart quotes etc.}
\usage{
fixMS(text, all = FALSE)
}
\arguments{
\item{text}{Text to be parsed}

\item{all}{Should all non-ascii characters be removed (default = FALSE)}
}
\description{
Replace Windows smart quotes etc.
}
