% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{flip}
\alias{flip}
\title{Flip the DT table to put Function, Variable, or Group by on top}
\usage{
flip(expl, top = "fun")
}
\arguments{
\item{expl}{Return value from \code{\link{explore}}}

\item{top}{The variable (type) to display at the top of the table ("fun" for Function, "var" for Variable, and "byvar" for Group by. "fun" is the default}
}
\description{
Flip the DT table to put Function, Variable, or Group by on top
}
\details{
See \url{https://radiant-rstats.github.io/docs/data/explore.html} for an example in Radiant
}
\examples{
explore(diamonds, "price:x", top = "var") \%>\% summary()
explore(diamonds, "price", byvar = "cut", fun = c("n_obs", "skew"), top = "byvar") \%>\% summary()

}
\seealso{
\code{\link{explore}} to calculate summaries

\code{\link{summary.explore}} to show summaries

\code{\link{dtab.explore}} to create the DT table
}
