% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{get_class}
\alias{get_class}
\title{Get variable class}
\usage{
get_class(dat)
}
\arguments{
\item{dat}{Dataset to evaluate}
}
\value{
Vector with class information for each variable
}
\description{
Get variable class
}
\details{
Get variable class information for each column in a data.frame
}
\examples{
get_class(mtcars)
}
