% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{ggplotly}
\alias{ggplotly}
\title{Work around to avoid (harmless) messages from ggplotly}
\usage{
ggplotly(...)
}
\arguments{
\item{...}{Arguments to pass to the \code{\link[plotly]{ggplotly}} function in the plotly package}
}
\description{
Work around to avoid (harmless) messages from ggplotly
}
\seealso{
See the \code{\link[plotly]{ggplotly}} function in the plotly package for details (?plotly::ggplotly)
}
