% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{glue_data}
\alias{glue_data}
\title{Exporting glue_data from glue}
\description{
Exporting glue_data from glue
}
\details{
See \code{\link[glue:glue]{glue::glue_data()}} in the \code{glue} package for more details
}
