% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{indexr}
\alias{indexr}
\title{Find index corrected for missing values and filters}
\usage{
indexr(dataset, vars = "", filt = "", arr = "", rows = NULL, cmd = "")
}
\arguments{
\item{dataset}{Dataset}

\item{vars}{Variables to select}

\item{filt}{Data filter}

\item{arr}{Expression to arrange (sort) the data on (e.g., "color, desc(price)")}

\item{rows}{Selected rows}

\item{cmd}{A command used to customize the data}
}
\description{
Find index corrected for missing values and filters
}
