% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{is_double}
\alias{is_double}
\title{Is input a double (and not a date type)?}
\usage{
is_double(x)
}
\arguments{
\item{x}{Input}
}
\value{
TRUE if double and not a type of date, else FALSE
}
\description{
Is input a double (and not a date type)?
}
