% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{is_numeric}
\alias{is_numeric}
\title{Is input numeric (and not a date type)?}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{Input}
}
\value{
TRUE if double and not a type of date, else FALSE
}
\description{
Is input numeric (and not a date type)?
}
