% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{iterms}
\alias{iterms}
\title{Create a vector of interaction terms for linear and logistic regression}
\usage{
iterms(vars, nway = 2, sep = ":")
}
\arguments{
\item{vars}{Variableslabels to use}

\item{nway}{2-way (2) or 3-way (3) interaction labels to create}

\item{sep}{Separator to use between variable names (e.g., :)}
}
\value{
Character vector of interaction term labels
}
\description{
Create a vector of interaction terms for linear and logistic regression
}
\examples{
paste0("var", 1:3) \%>\% iterms(2)
paste0("var", 1:3) \%>\% iterms(3)
paste0("var", 1:3) \%>\% iterms(2, sep = ".")

}
