% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{launch}
\alias{launch}
\title{Launch radiant apps in default browser or Rstudio viewer}
\usage{
launch(package = "radiant.data", run = "browser")
}
\arguments{
\item{package}{Radiant package to start. One of "radiant.data", "radiant.design", "radiant.basics", "radiant.model", "radiant.multivariate", "radiant"}

\item{run}{Run radiant app in an external browser ("browser") or in the Rstudio viewer ("viewer")}
}
\description{
Launch radiant apps in default browser or Rstudio viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
launch()
launch("viewer")
}

}
