% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{loadr}
\alias{loadr}
\title{Load an rds, rda, or csv file and add it to the radiant data list (r_data) if available}
\usage{
loadr(file, objname = "", rlist = TRUE)
}
\arguments{
\item{file}{File name and path as a string. Extension must be either rds, rda, or csv}

\item{objname}{Name to use for the data frame. Defaults to the file name}

\item{rlist}{If TRUE, uses "r_data" list to store the data.frame. If FALSE, loads data.frame into calling environment}
}
\value{
Data frame in r_data or in the calling enviroment
}
\description{
Load an rds, rda, or csv file and add it to the radiant data list (r_data) if available
}
