% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{make_arrange_cmd}
\alias{make_arrange_cmd}
\title{Generate arrange commands from user input}
\usage{
make_arrange_cmd(expr, dataset = "")
}
\arguments{
\item{expr}{Expression to use arrange rows from the specified dataset}

\item{dataset}{String with dataset name}
}
\value{
Arrange command
}
\description{
Generate arrange commands from user input
}
\details{
Form arrange command from user input
}
