% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_vec}
\alias{make_vec}
\title{Convert a string of numbers into a vector}
\usage{
make_vec(x)
}
\arguments{
\item{x}{A string of numbers that may include fractions}
}
\description{
Convert a string of numbers into a vector
}
\examples{
make_vec("1 2 4")
make_vec("1/2 2/3 4/5")
make_vec(0.1)
}
