% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{me}
\alias{me}
\title{Margin of error}
\usage{
me(x, conf_lev = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{conf_lev}{Confidence level. The default is 0.95}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Margin of error
}
\description{
Margin of error
}
\examples{
me(rnorm(100))

}
