% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mean_rm}
\alias{mean_rm}
\title{Mean with na.rm = TRUE}
\usage{
mean_rm(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Mean value
}
\description{
Mean with na.rm = TRUE
}
\examples{
mean_rm(runif (100))

}

