% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{median_rm}
\alias{median_rm}
\title{Median with na.rm = TRUE}
\usage{
median_rm(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Median value
}
\description{
Median with na.rm = TRUE
}
\examples{
median_rm(runif (100))

}
