% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{min_rm}
\alias{min_rm}
\title{Min with na.rm = TRUE}
\usage{
min_rm(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Minimum value
}
\description{
Min with na.rm = TRUE
}
\examples{
min_rm(runif (100))

}

