% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{modal}
\alias{modal}
\title{Calculate the mode (modal value) and return a label}
\usage{
modal(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\description{
Calculate the mode (modal value) and return a label
}
\details{
From https://www.tutorialspoint.com/r/r_mean_median_mode.htm
}
\examples{
modal(c("a", "b", "b"))
modal(c(1:10, 5))
modal(as.factor(c(letters, "b")))
modal(runif(100) > 0.5)

}
