% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{month}
\alias{month}
\title{Add ordered argument to lubridate::month}
\usage{
month(x, label = FALSE, abbr = TRUE, ordered = FALSE)
}
\arguments{
\item{x}{Input date vector}

\item{label}{Month as label (TRUE, FALSE)}

\item{abbr}{Abbreviate label (TRUE, FALSE)}

\item{ordered}{Order factor (TRUE, FALSE)}
}
\description{
Add ordered argument to lubridate::month
}
\seealso{
See the \code{\link[lubridate]{month}} function in the lubridate package for additional details
}
