% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mutate_if_tmp}
\alias{mutate_if_tmp}
\title{Temporary fix for mutate_if when the predicate is false for all columns}
\usage{
mutate_if_tmp(.tbl, .predicate, .funs, ...)
}
\arguments{
\item{.tbl}{Data frame}

\item{.predicate}{Predicate}

\item{.funs}{Function(s) to apply (e.g., funs(log))}

\item{...}{Additional arguments}
}
\description{
Temporary fix for mutate_if when the predicate is false for all columns
}
\details{
See https://github.com/tidyverse/dplyr/issues/2617
}
