% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{n_obs}
\alias{n_obs}
\title{Number of observations}
\usage{
n_obs(x, ...)
}
\arguments{
\item{x}{Input variable}

\item{...}{Additional arguments}
}
\value{
number of observations
}
\description{
Number of observations
}
\examples{
n_obs(c("a", "b", NA))

}
