% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p025}
\alias{p025}
\title{2.5th percentile}
\usage{
p025(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
2.5th percentile
}
\description{
2.5th percentile
}
\examples{
p025(rnorm(100))

}
