% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p975}
\alias{p975}
\title{97.5th percentile}
\usage{
p975(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
97.5th percentile
}
\description{
97.5th percentile
}
\examples{
p975(rnorm(100))

}
