% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{p01}
\alias{p01}
\alias{p025}
\alias{p05}
\alias{p10}
\alias{p25}
\alias{p75}
\alias{p90}
\alias{p95}
\alias{p975}
\alias{p99}
\alias{mode}
\title{Calculate percentiles}
\usage{
p01(x, na.rm = TRUE)

p025(x, na.rm = TRUE)

p05(x, na.rm = TRUE)

p10(x, na.rm = TRUE)

p25(x, na.rm = TRUE)

p75(x, na.rm = TRUE)

p90(x, na.rm = TRUE)

p95(x, na.rm = TRUE)

p975(x, na.rm = TRUE)

p99(x, na.rm = TRUE)

mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numer}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\description{
Calculate percentiles

Calculate the mode and return a label
}
\examples{
p01(0:100)

mode(c("a", "b", "b"))
mode(c(1:10, 5))
mode(as.factor(c(letters, "b")))
mode(runif(100) > 0.5)

}
