% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{plot.character}
\alias{plot.character}
\title{Don't try to plot strings}
\usage{
\method{plot}{character}(x, ...)
}
\arguments{
\item{x}{A character returned from a function}

\item{...}{Any additional arguments}
}
\description{
Don't try to plot strings
}
