% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr.R
\name{plot.pivotr}
\alias{plot.pivotr}
\title{Plot method for the pivotr function}
\usage{
\method{plot}{pivotr}(x, type = "dodge", perc = FALSE, flip = FALSE,
  fillcol = "blue", ...)
}
\arguments{
\item{x}{Return value from \code{\link{pivotr}}}

\item{type}{Plot type to use ("fill" or "dodge" (default))}

\item{perc}{Use percentage on the y-axis}

\item{flip}{Flip the axes in a plot (FALSE or TRUE)}

\item{fillcol}{Fill color for bar-plot when only one categorical variable has been selected (default is "blue")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the pivotr function
}
\details{
See \url{https://radiant-rstats.github.io/docs/data/pivotr} for an example in Radiant
}
\examples{
pivotr("diamonds", cvars = "cut") \%>\% plot
pivotr("diamonds", cvars = c("cut","clarity")) \%>\% plot
pivotr("diamonds", cvars = c("cut","clarity","color")) \%>\% plot

}
\seealso{
\code{\link{pivotr}} to generate summaries

\code{\link{summary.pivotr}} to show summaries
}
