% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{qscatter}
\alias{qscatter}
\title{Create a qscatter plot similar to Stata}
\usage{
qscatter(dataset, xvar, yvar, lev = "", fun = "mean", bins = 20)
}
\arguments{
\item{dataset}{Data to plot (data.frame or tibble)}

\item{xvar}{Character indicating the variable to display along the X-axis of the plot}

\item{yvar}{Character indicating the variable to display along the Y-axis of the plot}

\item{lev}{Level in yvar to use if yvar is of type character of factor. If lev is empty then the first level is used}

\item{fun}{Summary measure to apply to both the x and y variable}

\item{bins}{Number of bins to use}
}
\description{
Create a qscatter plot similar to Stata
}
\examples{
qscatter(diamonds, "price", "carat")
qscatter(titanic, "age", "survived")

}
