% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{qterms}
\alias{qterms}
\title{Create a vector of quadratic and cubed terms for use in linear and logist regression}
\usage{
qterms(vars, nway)
}
\arguments{
\item{vars}{Variables labels to use}

\item{nway}{2-way (2) or 3-way (3) interactions labels to create}
}
\value{
Character vector of regression term labels
}
\description{
Create a vector of quadratic and cubed terms for use in linear and logist regression
}
\examples{
qterms(c("a", "b"), 3)
qterms(c("a", "b"), 2)

}
