% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.base}
\alias{radiant.base}
\title{Launch Radiant in the default browser}
\usage{
radiant.base()
}
\description{
Launch Radiant in the default browser
}
\details{
See \url{http://vnijs.github.io/radiant} for documentation and tutorials
}

