% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.data_url}
\alias{radiant.data_url}
\title{Start radiant.data app but do not open a browser}
\usage{
radiant.data_url(state, ...)
}
\arguments{
\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Start radiant.data app but do not open a browser
}
\examples{
\dontrun{
radiant.data_url()
}
}
