% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.data_viewer}
\alias{radiant.data_viewer}
\title{Launch the radiant.data app in the Rstudio viewer}
\usage{
radiant.data_viewer(state, ...)
}
\arguments{
\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch the radiant.data app in the Rstudio viewer
}
\examples{
\dontrun{
radiant.data_viewer()
}
}
