% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.data_window}
\alias{radiant.data_window}
\title{Launch the radiant.data app in an Rstudio window}
\usage{
radiant.data_window()
}
\description{
Launch the radiant.data app in an Rstudio window
}
\examples{
\dontrun{
radiant.data_window()
}
}
