% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{read_files}
\alias{read_files}
\title{Return code to read file at specified path. Will open a file browser of no path is provided}
\usage{
read_files(path, type = "r", to = "", clipboard = TRUE, radiant = FALSE)
}
\arguments{
\item{path}{Path to file. If empty, a file browser will be opened}

\item{type}{Code for "rmd" or "r"}

\item{to}{Object name to use for object. If empty, will use file name to derive an appropriate object name}

\item{clipboard}{Return code to clipboard (not available on Linux)}

\item{radiant}{Should returned code be formatted for use with other code generated by Radiant}
}
\description{
Return code to read file at specified path. Will open a file browser of no path is provided
}
