% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{theme_version}
\alias{bs_theme}
\alias{wrap_plots}
\alias{plot_annotation}
\alias{writePNG}
\alias{glue}
\alias{glue_data}
\alias{glue_collapse}
\alias{knit_print}
\alias{rownames_to_column}
\alias{tibble}
\alias{as_tibble}
\alias{tidy}
\alias{glance}
\alias{kurtosi}
\alias{skew}
\alias{date}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{glance}}, \code{\link[broom:reexports]{tidy}}}

  \item{bslib}{\code{\link[bslib]{bs_theme}}, \code{\link[bslib]{theme_version}}}

  \item{glue}{\code{\link[glue]{glue}}, \code{\link[glue]{glue_collapse}}, \code{\link[glue:glue]{glue_data}}}

  \item{knitr}{\code{\link[knitr]{knit_print}}}

  \item{lubridate}{\code{\link[lubridate]{date}}}

  \item{patchwork}{\code{\link[patchwork]{plot_annotation}}, \code{\link[patchwork]{wrap_plots}}}

  \item{png}{\code{\link[png]{writePNG}}}

  \item{psych}{\code{\link[psych:skew]{kurtosi}}, \code{\link[psych]{skew}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble:rownames]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

