% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{register}
\alias{register}
\title{Register a data.frame in the datasetlist in Radiant}
\usage{
register(new, org = "", descr = "")
}
\arguments{
\item{new}{Name of the new dataset}

\item{org}{Name of the original data if a (working) copy is being made}

\item{descr}{Dataset description}
}
\description{
Register a data.frame in the datasetlist in Radiant
}
