% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render.character}
\alias{render.character}
\title{Method to render rmarkdown documents}
\usage{
\method{render}{character}(object, ...)
}
\arguments{
\item{object}{File path to an R-markdown file}

\item{...}{Additional arguments passed on to rmarkdown::render}
}
\description{
Method to render rmarkdown documents
}
