% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render.datatables}
\alias{render.datatables}
\title{Method to render DT tabels}
\usage{
\method{render}{datatables}(object, ...)
}
\arguments{
\item{object}{DT table plot}

\item{...}{Additional arguments}
}
\description{
Method to render DT tabels
}
