% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render.plotly}
\alias{render.plotly}
\title{Method to render plotly plots}
\usage{
\method{render}{plotly}(object, shiny = shiny::getDefaultReactiveDomain(), ...)
}
\arguments{
\item{object}{plotly object}

\item{shiny}{Check if function is called from a shiny application}

\item{...}{Additional arguments}
}
\description{
Method to render plotly plots
}
