% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render}
\alias{render}
\title{Method to render htmlwidgets}
\usage{
render(object, ...)
}
\arguments{
\item{object}{Object of relevant class to render}

\item{...}{Additional arguments}
}
\description{
Method to render htmlwidgets
}

