% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render.shiny.render.function}
\alias{render.shiny.render.function}
\title{Method to avoid re-rendering a shiny.render.function}
\usage{
\method{render}{shiny.render.function}(object, ...)
}
\arguments{
\item{object}{Shiny render function}

\item{...}{Additional arguments}
}
\description{
Method to avoid re-rendering a shiny.render.function
}
