% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{round_df}
\alias{round_df}
\title{Round doubles in a data.frame to a specified number of decimal places}
\usage{
round_df(tbl, dec = 3)
}
\arguments{
\item{tbl}{Data frame}

\item{dec}{Number of decimals to show}
}
\value{
Data frame with rounded doubles
}
\description{
Round doubles in a data.frame to a specified number of decimal places
}
\examples{
data.frame(x = as.factor(c("a", "b")), y = c(1L, 2L), z = c(-0.0005, 3.1)) \%>\%
  round_df(dec = 2)

}
