% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{save_clip}
\alias{save_clip}
\title{Save data.frame or tibble to clipboard on Windows or macOS}
\usage{
save_clip(dataset)
}
\arguments{
\item{dataset}{Dataset to push to clipboard}
}
\description{
Save data.frame or tibble to clipboard on Windows or macOS
}
\details{
See \url{https://radiant-rstats.github.io/docs/data/manage.html} for an example in Radiant
}
