% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{sdp_rm}
\alias{sdp_rm}
\title{Standard deviation for the population na.rm = TRUE}
\usage{
sdp_rm(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Standard deviation for the population
}
\description{
Standard deviation for the population na.rm = TRUE
}
\examples{
sdp_rm(rnorm(100))

}

