% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{sdpop}
\alias{sdpop}
\title{Standard deviation for the population}
\usage{
sdpop(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Standard deviation for the population
}
\description{
Standard deviation for the population
}
\examples{
sdpop(rnorm(100))

}
