% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{sdprop}
\alias{sdprop}
\title{Standard deviation for proportion}
\usage{
sdprop(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Standard deviation for proportion
}
\description{
Standard deviation for proportion
}
\examples{
sdprop(c(rep(1L, 10), rep(0L, 10)))

}

