% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{Search}
\alias{Search}
\title{Search for a string in all columns of a data.frame}
\usage{
Search(pattern, df, ignore.case = TRUE, fixed = FALSE)
}
\arguments{
\item{pattern}{String to match}

\item{df}{Data.frame to search}

\item{ignore.case}{Should search be case sensitive or not (default is FALSE)}

\item{fixed}{Allow regular expersions or not (default is FALSE)}
}
\description{
Search for a string in all columns of a data.frame
}
\details{
See \url{http://radiant-rstats.github.io/docs/data/view.html} for an example in Radiant
}
\seealso{
See \code{\link{grepl}} for a more detailed description of the function arguments
}

