% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{search_data}
\alias{search_data}
\title{Search for a pattern in all columns of a data.frame}
\usage{
search_data(dataset, pattern, ignore.case = TRUE, fixed = FALSE)
}
\arguments{
\item{dataset}{Data.frame to search}

\item{pattern}{String to match}

\item{ignore.case}{Should search be case sensitive or not (default is FALSE)}

\item{fixed}{Allow regular expressions or not (default is FALSE)}
}
\description{
Search for a pattern in all columns of a data.frame
}
\examples{
publishers \%>\% filter(search_data(., "^m"))
}
\seealso{
See \code{\link{grepl}} for a detailed description of the function arguments
}
