% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{seprop}
\alias{seprop}
\title{Standard error for proportion}
\usage{
seprop(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Standard error for proportion
}
\description{
Standard error for proportion
}
\examples{
seprop(c(rep(1L, 10), rep(0L, 10)))

}
