% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{sig_stars}
\alias{sig_stars}
\title{Add stars based on p.values}
\usage{
sig_stars(pval)
}
\arguments{
\item{pval}{Vector of p-values}
}
\value{
A vector of stars
}
\description{
Add stars based on p.values
}
\examples{
sig_stars(c(.0009, .049, .009, .4, .09))
}
