% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{standardize}
\alias{standardize}
\title{Standardize}
\usage{
standardize(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
If x is a numeric variable return center(x) / mean(x)
}
\description{
Standardize
}
