% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{store.character}
\alias{store.character}
\title{Method for error messages that a user tries to store}
\usage{
\method{store}{character}(object, ...)
}
\arguments{
\item{object}{Object of type character}

\item{...}{Additional arguments}
}
\description{
Method for error messages that a user tries to store
}

