% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{store.data.frame}
\alias{store.data.frame}
\title{Store method for the Data > View tab}
\usage{
\method{store}{data.frame}(object, new = "", org = "",
  envir = parent.frame(), ...)
}
\arguments{
\item{object}{Filtered data frame from the Data > View tab}

\item{new}{Name of the new dataset}

\item{org}{Name of the original data}

\item{envir}{Environment to assign 'new' dataset (optional). Used if 'new' is specified but an r_data list is not available}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store method for the Data > View tab
}
\details{
Store data frame in Radiant r_data list if available
}
