% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{store.explore}
\alias{store.explore}
\title{Deprecated: Store method for the explore function}
\usage{
\method{store}{explore}(dataset, object, name, ...)
}
\arguments{
\item{dataset}{Dataset}

\item{object}{Return value from \code{\link{explore}}}

\item{name}{Name to assign to the dataset}

\item{...}{further arguments passed to or from other methods}
}
\description{
Deprecated: Store method for the explore function
}
\details{
Return the summarized data. See \url{https://radiant-rstats.github.io/docs/data/explore.html} for an example in Radiant
}
\seealso{
\code{\link{explore}} to generate summaries
}
