% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr.R
\name{store.pivotr}
\alias{store.pivotr}
\title{Deprecated: Store method for the pivotr function}
\usage{
\method{store}{pivotr}(dataset, object, name, ...)
}
\arguments{
\item{dataset}{Dataset}

\item{object}{Return value from \code{\link{pivotr}}}

\item{name}{Name to assign to the dataset}

\item{...}{further arguments passed to or from other methods}
}
\description{
Deprecated: Store method for the pivotr function
}
\details{
Return the summarized data. See \url{https://radiant-rstats.github.io/docs/data/pivotr.html} for an example in Radiant
}
\seealso{
\code{\link{pivotr}} to generate summaries
}
