% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{store.view}
\alias{store.view}
\title{Store method for the Data > View tab}
\usage{
\method{store}{view}(object, name1, name2, ...)
}
\arguments{
\item{object}{Filtered data frame from the Data > View tab}

\item{name1}{Name of the original data}

\item{name2}{Name of the new dataset}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store method for the Data > View tab
}
\details{
Apply all filters and searches. See \url{http://radiant-rstats.github.io/docs/data/view.html} for an example in Radiant
}

