% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{summary.explore}
\alias{summary.explore}
\title{Summary method for the explore function}
\usage{
\method{summary}{explore}(object, dec = 3, ...)
}
\arguments{
\item{object}{Return value from \code{\link{explore}}}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the explore function
}
\details{
See \url{https://radiant-rstats.github.io/docs/data/explore.html} for an example in Radiant
}
\examples{
result <- explore(diamonds, "price:x")
summary(result)
result <- explore(diamonds, "price", byvar = "cut", fun = c("n_obs", "skew"))
summary(result)
explore(diamonds, "price:x", byvar = "color") \%>\% summary()

}
\seealso{
\code{\link{explore}} to generate summaries
}
