% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{table2data}
\alias{table2data}
\title{Create data.frame from a table}
\usage{
table2data(dataset, freq = tail(colnames(dataset), 1))
}
\arguments{
\item{dataset}{Data.frame}

\item{freq}{Column name with frequency information}
}
\description{
Create data.frame from a table
}
\examples{
data.frame(price = c("$200", "$300"), sale = c(10, 2)) \%>\% table2data()

}
