% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{var_rm}
\alias{var_rm}
\title{Variance with na.rm = TRUE}
\usage{
var_rm(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Variance
}
\description{
Variance with na.rm = TRUE
}
\examples{
var_rm(rnorm(100))

}

