% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{varpop}
\alias{varpop}
\title{Variance for the population}
\usage{
varpop(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Variance for the population
}
\description{
Variance for the population
}
\examples{
varpop(rnorm(100))

}
