% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{varprop}
\alias{varprop}
\title{Variance for proportion}
\usage{
varprop(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
Variance for proportion
}
\description{
Variance for proportion
}
\examples{
varprop(c(rep(1L, 10), rep(0L, 10)))

}
