% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_data}
\alias{view_data}
\title{View data in a shiny-app}
\usage{
view_data(dataset, vars = "", filt = "", rows = NULL, na.rm = FALSE,
  dec = 3)
}
\arguments{
\item{dataset}{Data.frame or name of the dataframe to view}

\item{vars}{Variables to show (default is all)}

\item{filt}{Filter to apply to the specified dataset}

\item{rows}{Select rows in the specified dataset}

\item{na.rm}{Remove rows with missing values (default is FALSE)}

\item{dec}{Number of decimals to show}
}
\description{
View data in a shiny-app
}
\details{
View, search, sort, etc. your data
}
\examples{
\dontrun{
view_data(mtcars)
}

}
\seealso{
See \code{\link{get_data}} and \code{\link{filter_data}}
}
