% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wday}
\alias{wday}
\title{Add ordered argument to lubridate::wday}
\usage{
wday(x, label = FALSE, abbr = TRUE, ordered = FALSE)
}
\arguments{
\item{x}{Input date vector}

\item{label}{Weekday as label (TRUE, FALSE)}

\item{abbr}{Abbreviate label (TRUE, FALSE)}

\item{ordered}{Order factor (TRUE, FALSE)}
}
\description{
Add ordered argument to lubridate::wday
}
\seealso{
See the \code{\link[lubridate:day]{lubridate::wday()}} function in the lubridate package for additional details
}
