% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{which.pmin}
\alias{which.pmin}
\title{Index of the minimum per row}
\usage{
which.pmin(...)
}
\arguments{
\item{...}{Numeric or character vectors of the same length}
}
\value{
Vector of rankings
}
\description{
Index of the minimum per row
}
\details{
Determine the index of the minimum of the input vectors per row. Extension of \code{which.min}
}
\examples{
which.pmin(1:10, 10:1)
which.pmin(2, 10:1)
which.pmin(mtcars)
}
\seealso{
See also \code{\link{which.min}} and \code{\link{which.pmax}}
}
