% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{write_feather}
\alias{write_feather}
\title{Workaround to add description using feather::write_feather}
\usage{
write_feather(x, path, description = attr(x, "description"))
}
\arguments{
\item{x}{A data frame to write to disk}

\item{path}{Path to feather file}

\item{description}{Data description}
}
\description{
Workaround to add description using feather::write_feather
}
