% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{write_parquet}
\alias{write_parquet}
\title{Workaround to store description file together with a parquet data file}
\usage{
write_parquet(x, file, description = attr(x, "description"))
}
\arguments{
\item{x}{A data frame to write to disk}

\item{file}{Path to store parquet file}

\item{description}{Data description}
}
\description{
Workaround to store description file together with a parquet data file
}
