% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{describe}
\alias{describe}
\title{Show dataset description}
\usage{
describe(dataset, envir = parent.frame())
}
\arguments{
\item{dataset}{Dataset with "description" attribute}

\item{envir}{Environment to extract data from}
}
\description{
Show dataset description
}
\details{
Show dataset description, if available, in html form in Rstudio viewer or the default browser. The description should be in markdown format, attached to a data.frame as an attribute with the name "description"
}
