% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{describe}
\alias{describe}
\title{Show dataset description, if available, in html form in Rstudio viewer or default browser}
\usage{
describe(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\description{
Show dataset description, if available, in html form in Rstudio viewer or default browser
}
