% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{dtab.explore}
\alias{dtab.explore}
\title{Make a tabel of summary statistics in DT}
\usage{
\method{dtab}{explore}(object, dec = 3, searchCols = NULL, order = NULL,
  pageLength = NULL, ...)
}
\arguments{
\item{object}{Return value from \code{\link{explore}}}

\item{dec}{Number of decimals to show}

\item{searchCols}{Column search and filter. Used to save and restore state}

\item{order}{Column sorting. Used to save and restore state}

\item{pageLength}{Page length. Used to save and restore state}

\item{...}{further arguments passed to or from other methods}
}
\description{
Make a tabel of summary statistics in DT
}
\details{
See \url{http://radiant-rstats.github.io/docs/data/explore.html} for an example in Radiant
}
\examples{
tab <- explore("diamonds", "price:x") \%>\% dtab
tab <- explore("diamonds", "price", byvar = "cut", fun = c("length", "skew"), top = "byvar") \%>\%
  dtab

}
\seealso{
\code{\link{pivotr}} to create the pivot-table using dplyr

\code{\link{summary.pivotr}} to print a plain text table
}

