% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{dtab}
\alias{dtab}
\title{Method to create datatables}
\usage{
dtab(object, ...)
}
\arguments{
\item{object}{Object of relevant class to render}

\item{...}{Additional arguments}
}
\description{
Method to create datatables
}
\seealso{
See \code{\link{dtab.explore}} to create the an interactivce table from an \code{\link{explore}} object

See \code{\link{dtab.pivotr}} to create the an interactivce table from a \code{\link{pivotr}} object

See \code{\link{dtab.data.frame}} to create an interactive table from a data.frame
}
