% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{formatdf}
\alias{formatdf}
\title{Format a data.frame with a specified number of decimal places}
\usage{
formatdf(tbl, dec = 3, perc = FALSE, mark = "")
}
\arguments{
\item{tbl}{Data.frame}

\item{dec}{Number of decimal places}

\item{perc}{Display numbers as percentages (TRUE or FALSE)}

\item{mark}{Thousand separator}
}
\value{
Data.frame for printing
}
\description{
Format a data.frame with a specified number of decimal places
}
\examples{
data.frame(x = c("a","b"), y = c(1L, 2L), z = c(-0.0005, 3)) \%>\%
  formatdf(dec = 3)
data.frame(x = c(1L, 2L), y = c(0.05, 0.8)) \%>\%
  formatdf(dec = 2, perc = TRUE)

}

