% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{formatnr}
\alias{formatnr}
\title{Format a number with a specified number of decimal places, thousand sep, and a symbol}
\usage{
formatnr(x, sym = "", dec = 2, perc = FALSE, mark = ",")
}
\arguments{
\item{x}{Number or vector}

\item{sym}{Symbol to use}

\item{dec}{Number of decimal places}

\item{perc}{Display number as a percentage}

\item{mark}{Thousand separator}
}
\value{
Character (vector) in the desired format
}
\description{
Format a number with a specified number of decimal places, thousand sep, and a symbol
}
\examples{
formatnr(2000, "$")
formatnr(2000, dec = 4)
formatnr(.05, perc = TRUE)
formatnr(c(.1, .99), perc = TRUE)
formatnr(data.frame(a = c(.1, .99)), perc = TRUE)
formatnr(data.frame(a = 1000), sym = "$", dec = 0)

}

