% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{ggplotly}
\alias{ggplotly}
\title{Workaround to avoid messages from ggplotly}
\usage{
ggplotly(...)
}
\arguments{
\item{...}{Arguments to pass to plotly::ggplotlyy}
}
\description{
Workaround to avoid messages from ggplotly
}
\seealso{
See the \code{\link[plotly]{ggplotly}} in the plotly package for details (?plotly::ggplotly)
}
