% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{glue_collapse}
\alias{glue_collapse}
\title{Exporting collapse from glue as glue_collapse}
\usage{
glue_collapse(...)
}
\arguments{
\item{...}{arguments passed to \code{glue::glue_collapse}}
}
\description{
Exporting collapse from glue as glue_collapse
}
\details{
See \code{\link[glue]{collapse}} in the \code{glue} package for more details
}
