% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{load_clip}
\alias{load_clip}
\title{Load data through clipboard on Windows or macOS}
\usage{
load_clip(delim = "\\t", text, suppress = TRUE)
}
\arguments{
\item{delim}{Delimiter to use (tab is the default)}

\item{text}{Text input to convert to table}

\item{suppress}{Suppress warnings}
}
\description{
Load data through clipboard on Windows or macOS
}
\details{
See \url{https://radiant-rstats.github.io/docs/data/manage.html} for an example in Radiant
}
