% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{parse_path}
\alias{parse_path}
\title{Parse path into useful components (used by read_files function)}
\usage{
parse_path(path, chr = "\\"", pdir = getOption("radiant.project_dir",
  default = rstudioapi::getActiveProject()))
}
\arguments{
\item{path}{Path to be parsed}

\item{chr}{Character to wrap around path for display}

\item{pdir}{Project directory if available}
}
\description{
Parse path into useful components (used by read_files function)
}
