% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{qterms}
\alias{qterms}
\title{Create a vector of quadratic and cubed terms for use in linear and logist regression}
\usage{
qterms(vars, nway = 2)
}
\arguments{
\item{vars}{Variables labels to use}

\item{nway}{quadratic (2) or cubic (3) term labels to create}
}
\value{
Character vector of (regression) term labels
}
\description{
Create a vector of quadratic and cubed terms for use in linear and logist regression
}
\examples{
qterms(c("a", "b"), 3)
qterms(c("a", "b"), 2)

}
