% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{radiant.data-deprecated}
\alias{dfprint}
\alias{nrprint}
\alias{radiant.data-deprecated}
\alias{radiant.data-deprecated-package}
\title{Deprecated function(s) in the radiant.data package}
\usage{
dfprint(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
radiant. They will eventually be  removed.
}
\section{Details}{

\tabular{rl}{
  \code{dfprint} is now a synonym for \code{\link{formatdf}}\cr
  \code{nrprint} is now a synonym for \code{\link{formatnr}}\cr
  \code{varp_rm} is now a synonym for \code{\link{varpop}}\cr
  \code{sdp_rm} is now a synonym for \code{\link{sdpop}}\cr
}
}

