% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{radiant.data-deprecated}
\alias{radiant.data-deprecated}
\alias{mean_rm}
\alias{mean_rmn}
\alias{radiant.data-deprecated-package}
\title{Deprecated function(s) in the radiant.data package}
\usage{
mean_rm(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
radiant but will be removed
}
\section{Details}{

\tabular{rl}{
  Replace \code{mean_rm} by \code{\link{mean}}\cr
  Replace \code{median_rm} by \code{\link{median}}\cr
  Replace \code{min_rm} by \code{\link{min}}\cr
  Replace \code{max_rm} by \code{\link{max}}\cr
  Replace \code{sd_rm} by \code{\link{sd}}\cr
  Replace \code{var_rm} by \code{\link{var}}\cr
  Replace \code{sum_rm} by \code{\link{sum}}\cr
}
}

