% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{radiant.data-deprecated}
\alias{radiant.data-deprecated}
\alias{mean_rm}
\alias{median_rm}
\alias{min_rm}
\alias{max_rm}
\alias{sd_rm}
\alias{var_rm}
\alias{sum_rm}
\alias{radiant.data-deprecated-package}
\title{Deprecated function(s) in the radiant.data package}
\usage{
mean_rm(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
radiant but will be removed
}
\section{Details}{

\itemize{
  \item Replace \code{mean_rm} by \code{\link{mean}}
  \item Replace \code{median_rm} by \code{\link{median}}
  \item Replace \code{min_rm} by \code{\link{min}}
  \item Replace \code{max_rm} by \code{\link{max}}
  \item Replace \code{sd_rm} by \code{\link{sd}}
  \item Replace \code{var_rm} by \code{\link{var}}
  \item Replace \code{sum_rm} by \code{\link{sum}}
}
}

