% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.data}
\alias{radiant.data}
\alias{radiant.data-package}
\title{radiant.data}
\usage{
radiant.data()
}
\description{
radiant.data

Launch the radiant.data app in the default web browser
}
\examples{
\dontrun{
radiant.data()
radiant.data("viewer")
}
}
