% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.data}
\alias{radiant.data}
\title{radiant.data}
\usage{
radiant.data(state, ...)
}
\arguments{
\item{state}{Path to statefile to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch the radiant.data app in the default web browser
}
\examples{
\dontrun{
radiant.data()
radiant.data("https://github.com/radiant-rstats/docs/raw/gh-pages/examples/demo-dvd-rnd.state.rda")
radiant.data("viewer")
}
}
