% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.data}
\alias{radiant.data}
\alias{radiant.data-package}
\alias{radiant.data}
\title{radiant.data}
\usage{
radiant.data(run = "browser")
}
\arguments{
\item{run}{Run radiant.data in an external browser ("browser") or in the Rstudio viewer ("viewer")}
}
\description{
radiant.data

Launch radiant.data in default browser or Rstudio Viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.data::radiant.data()
radiant.data::radiant.data("viewer")
}

}
