% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.data}
\alias{radiant.data}
\alias{radiant.data-package}
\title{radiant.data}
\usage{
radiant.data()
}
\description{
radiant.data

Launch Radiant in the default browser
}
\details{
See \url{http://vnijs.github.io/radiant} for documentation and tutorials
}

