% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.data_window}
\alias{radiant.data_window}
\title{Launch radiant.data in the Rstudio window}
\usage{
radiant.data_window()
}
\description{
Launch radiant.data in the Rstudio window
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.data_window()
}
}
