% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{read_files}
\alias{read_files}
\title{Generate code to read a file}
\usage{
read_files(path, pdir = "", type = "rmd", to = "", clipboard = TRUE,
  radiant = FALSE)
}
\arguments{
\item{path}{Path to file. If empty, a file browser will be opened}

\item{pdir}{Project dir}

\item{type}{Generate code for _Report > Rmd_ ("rmd") or _Report > R_ ("r")}

\item{to}{Name to use for object. If empty, will use file name to derive an object name}

\item{clipboard}{Return code to clipboard (not available on Linux)}

\item{radiant}{Should returned code be formatted for use with other code generated by Radiant?}
}
\description{
Generate code to read a file
}
\details{
Return code to read a file at the specified path. Will open a file browser if no path is provided
}
\examples{
if (interactive()) {
read_files(clipboard = FALSE)
}
}
