% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{wrap_plots}
\alias{plot_annotation}
\alias{glue}
\alias{glue_data}
\alias{glue_collapse}
\alias{knit_print}
\alias{rownames_to_column}
\alias{tibble}
\alias{as_tibble}
\alias{tidy}
\alias{glance}
\alias{kurtosi}
\alias{skew}
\alias{date}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{glance}}, \code{\link[broom]{tidy}}}

  \item{glue}{\code{\link[glue]{glue}}, \code{\link[glue]{glue_collapse}}, \code{\link[glue]{glue_data}}}

  \item{knitr}{\code{\link[knitr]{knit_print}}}

  \item{lubridate}{\code{\link[lubridate]{date}}}

  \item{patchwork}{\code{\link[patchwork]{plot_annotation}}, \code{\link[patchwork]{wrap_plots}}}

  \item{psych}{\code{\link[psych]{kurtosi}}, \code{\link[psych]{skew}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

