% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{register}
\alias{register}
\title{Register a data.frame or list in Radiant}
\usage{
register(new, org = "", descr = "",
  shiny = shiny::getDefaultReactiveDomain(), envir = r_data)
}
\arguments{
\item{new}{String containing the name of the data.frame to register}

\item{org}{Name of the original data.frame if a (working) copy is being made}

\item{descr}{Data description in markdown format}

\item{shiny}{Check if function is called from a shiny application}

\item{envir}{Environment to assign data to}
}
\description{
Register a data.frame or list in Radiant
}
