% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{register}
\alias{register}
\title{Register a data.frame or list Radiant}
\usage{
register(new, org = "", descr = "", env = r_data)
}
\arguments{
\item{new}{String containing the name of the data.frame or tibble to register}

\item{org}{Name of the original data.frame or tibble if a (working) copy is being made}

\item{descr}{Data description in markdown format}

\item{env}{Environment containing the data}
}
\description{
Register a data.frame or list Radiant
}
