% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{render.datatables}
\alias{render.datatables}
\title{Method to render DT tables}
\usage{
\method{render}{datatables}(object, shiny = shiny::getDefaultReactiveDomain(), ...)
}
\arguments{
\item{object}{DT table}

\item{shiny}{Check if function is called from a shiny application}

\item{...}{Additional arguments}
}
\description{
Method to render DT tables
}
