% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{save_clip}
\alias{save_clip}
\title{Save data to clipboard on Windows or macOS}
\usage{
save_clip(dataset)
}
\arguments{
\item{dataset}{Dataset to save to clipboard}
}
\description{
Save data to clipboard on Windows or macOS
}
\details{
Save a data.frame or tibble to the clipboard on Windows or macOS
}
\seealso{
See the \code{\link{load_clip}}
}
