% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{set_attr}
\alias{set_attr}
\title{Alias used to add an attribute (from github version of magrittr)}
\usage{
set_attr()
}
\description{
Alias used to add an attribute (from github version of magrittr)
}
\examples{
foo <- data.frame(price = 1:5) \%>\% set_attr("desc", "price set in experiment ...")

}

