% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{slice_data}
\alias{slice_data}
\title{Slice data with user-specified expression}
\usage{
slice_data(dataset, expr = NULL, drop = TRUE)
}
\arguments{
\item{dataset}{Data frame to slice}

\item{expr}{Expression to use select rows from the specified dataset}

\item{drop}{Drop unused factor levels after filtering (default is TRUE)}
}
\value{
Sliced data frame
}
\description{
Slice data with user-specified expression
}
\details{
Select only a slice of the data to work with
}
