% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{store.explore}
\alias{store.explore}
\title{Store method for the explore function}
\usage{
\method{store}{explore}(object, name, ...)
}
\arguments{
\item{object}{Return value from \code{\link{explore}}}

\item{name}{Name to assign to the dataset}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store method for the explore function
}
\details{
Add the summarized data to the r_data list in Radiant or return it. See \url{http://radiant-rstats.github.io/docs/data/explore.html} for an example in Radiant
}
\seealso{
\code{\link{explore}} to generate summaries
}

