% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr.R
\name{store.pivotr}
\alias{store.pivotr}
\title{Store method for the pivort function}
\usage{
\method{store}{pivotr}(object, name, ...)
}
\arguments{
\item{object}{Return value from \code{\link{pivotr}}}

\item{name}{Name to assign to the dataset}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store method for the pivort function
}
\details{
Add the summarized data to the r_data list in Radiant or return it. See \url{http://radiant-rstats.github.io/docs/data/pivotr.html} for an example in Radiant
}
\seealso{
\code{\link{pivotr}} to generate summaries
}
