% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{subplot}
\alias{subplot}
\title{Workaround to avoid messages from subplot}
\usage{
subplot(..., margin = 0.04)
}
\arguments{
\item{...}{Arguments to pass to plotly::subplot}

\item{margin}{Default margin to use between plots}
}
\description{
Workaround to avoid messages from subplot
}
\seealso{
See the \code{\link[plotly]{subplot}} in the plotly package for details (?plotly::subplot)
}
