% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{xtile}
\alias{xtile}
\title{Create quantiles}
\usage{
xtile(x, n = 5, rev = FALSE)
}
\arguments{
\item{x}{Numeric variable}

\item{n}{number of bins to create}

\item{rev}{Reverse the order of the xtiles}
}
\description{
Create quantiles
}
\details{
Approach used produces results most similar to Stata
}
\examples{
xtile(1:10,5)
xtile(1:10,5, rev = TRUE)

}
